refactor: move splitDashSeparatedOptionalParams to a new helper module
